# feat: Add Advanced Detection Algorithms (70% Attack Coverage)