… DETECTION_ALGORITHM_ANALYSIS.md
… NEW_DETECTORS_IMPLEMENTATION.md
… backend/src/services/detection/timing-variance-detector.service.ts
… backend/tests/services/detection/data-volume-detector.test.ts
… backend/tests/services/detection/permission-escalation-detector.test.ts
… backend/tests/services/detection/timing-variance-detector.test.ts